% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.r
\name{dict_from_string}
\alias{dict_from_string}
\title{Create a dictionary from a string format}
\usage{
dict_from_string(txt)
}
\arguments{
\item{txt}{A string in which each line is a query. If the line contains a
"=" sign, the left side becomes the label, and the right
side the query. See examples below.}
}
\value{
A query dictionary
}
\description{
This is a convenience function for typing queries in R.
See examples below for how to use the function.
}
\examples{

dict_from_string('
   Barack Obama = "(barack OR president) obama"
   spices = parsley OR sage OR rosemary OR thyme
')
}
