% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.r
\name{dict_join}
\alias{dict_join}
\alias{dict_left_join}
\alias{dict_right_join}
\alias{dict_full_join}
\title{Join two tables based on a character column of a dictionary or Lucene-like
Boolean queries}
\usage{
dict_join(x, y, by, mode = "left", ...)

dict_left_join(x, y, by, ...)

dict_right_join(x, y, by, ...)

dict_full_join(x, y, by, ...)
}
\arguments{
\item{x, y}{A pair of data frames, data frame extensions (e.g. a tibble)}

\item{by}{A character vector of variables to join by. To join on different
variables between `x` and `y`, use a named vector. For example, `by = c("a"
= "b")` will match `x$a` to `y$b`.}

\item{mode}{One of "left", "right" or "full".}

\item{...}{Handed to \link{search_query}.}
}
\value{
A data.table or tibble
}
\description{
Essentially dplyr's join operations (see \link[dplyr]{join}) but with
additional query features.
}
\examples{
df = data.frame(
  text = c('This is just a simple example', 'Simple is good'),
  doc_id = c(1, 2)
)
dict_join(df, dict, by = c("text" = "string"))
dict_left_join(df, dict, by = c("text" = "string"))
dict_right_join(df, dict, by = c("text" = "string"))
dict_full_join(df, dict, by = c("text" = "string"))
}
