% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{is_advanced}
\alias{is_advanced}
\title{Check if string is a simple dictionary term or advanced query}
\usage{
is_advanced(q)
}
\arguments{
\item{q}{a character vector}
}
\value{
A logical vector
}
\description{
Check if string is a simple dictionary term or advanced query
}
\examples{

queries = c('this OR that',    ## boolean, so advanced
            'this that',       ## space means OR, so advanced
            'single',          ## single term, so basic
            '<simple multiword str*>')  ## quoted, so basic

is_advanced(queries)
}
