% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\name{query_function}
\alias{query_function}
\alias{query_subset}
\alias{query_join}
\alias{query_aggregate}
\title{Two functions of sum1 and prod1}
\usage{
query_subset()

query_join()

query_aggregate()
}
\arguments{
\item{x}{=X}

\item{y}{=Y}
}
\value{
x*y (prod1) or x+y (sum1).
}
\description{
Two functions of sum1 and prod1
}
